Refectored scenario tests according to #Bug 240688: Scenarios has a class Top5Scenarios with 2 scenarios defined
