+++ test/Microsoft.ML.Tests/Scenarios/SentimentPredictionTests.cs
@@ -13,7 +13,7 @@ using Xunit;
 
 namespace Microsoft.ML.Scenarios
 {
-    public partial class Top5Scenarios
+    public partial class ScenariosTests
     {
         public const string SentimentDataPath = "wikipedia-detox-250-line-data.tsv";
         public const string SentimentTestPath = "wikipedia-detox-250-line-test.tsv";
